feat(index.html): add new presentation HTML file for AU-SSI 2025
Introduce a new HTML file to serve as the presentation
for the AU-SSI 2025 event titled "Surface growth by Le Chen."
The file includes various sections with slides, images, and
videos to illustrate concepts such as surface growth, the
Central Limit Theorem, and other related topics. The
presentation is designed to be interactive and informative,
utilizing the Reveal.js framework for slide transitions and
MathJax for mathematical expressions. This addition aims to
enhance the educational experience for attendees by providing
a visually engaging and comprehensive overview of the subject
matter.