docs(README.md): add group photo and documents section
Add a group photo from the session to provide a visual element
that captures the event. Include links to a thank you card and
program summary to offer additional resources and context about
the 2025 SSI program. These additions enhance the documentation
by providing more comprehensive information and a personal touch.